feat(functions): add checks for argument presence in functions, and correctly use $* and $@ vars

+++ lib/functions/system.sh
@@ -5,10 +5,20 @@ shutdn() {
 }
 
 kk() {
+  if [[ -z "$1" ]]; then
+    echo "Usage: kk <process-name>"
+    return 1
+  fi
+
   sudo killall "$1" -9
 }
 
 mnt() {
+  if [[ -z "$1" ]]; then
+    echo "Usage: mnt <device>"
+    return 1
+  fi
+
   sudo mount "/dev/$1" /mnt
 }
